+++ agentidentity-20250901/alibabacloud_agentidentity20250901/models/_create_oauth2_credential_provider_request.py
@@ -13,6 +13,7 @@ class CreateOAuth2CredentialProviderRequest(DaraModel):
         description: str = None,
         oauth2_credential_provider_name: str = None,
         oauth2_provider_config: main_models.OAuth2ProviderConfig = None,
+        oauth_type: str = None,
         token_vault_name: str = None,
     ):
         self.callback_url = callback_url
@@ -20,6 +21,7 @@ class CreateOAuth2CredentialProviderRequest(DaraModel):
         self.description = description
         self.oauth2_credential_provider_name = oauth2_credential_provider_name
         self.oauth2_provider_config = oauth2_provider_config
+        self.oauth_type = oauth_type
         self.token_vault_name = token_vault_name
 
     def validate(self):
@@ -46,6 +48,9 @@ class CreateOAuth2CredentialProviderRequest(DaraModel):
         if self.oauth2_provider_config is not None:
             result['OAuth2ProviderConfig'] = self.oauth2_provider_config.to_map()
 
+        if self.oauth_type is not None:
+            result['OAuthType'] = self.oauth_type
+
         if self.token_vault_name is not None:
             result['TokenVaultName'] = self.token_vault_name
 
@@ -68,6 +73,9 @@ class CreateOAuth2CredentialProviderRequest(DaraModel):
         if m.get('OAuth2ProviderConfig') is not None:
             temp_model = main_models.OAuth2ProviderConfig()
             self.oauth2_provider_config = temp_model.from_map(m.get('OAuth2ProviderConfig'))
+
+        if m.get('OAuthType') is not None:
+            self.oauth_type = m.get('OAuthType')
 
         if m.get('TokenVaultName') is not None:
             self.token_vault_name = m.get('TokenVaultName')