+++ agentidentity-20250901/alibabacloud_agentidentity20250901/models/_create_oauth2_credential_provider_shrink_request.py
@@ -12,6 +12,7 @@ class CreateOAuth2CredentialProviderShrinkRequest(DaraModel):
         description: str = None,
         oauth2_credential_provider_name: str = None,
         oauth2_provider_config_shrink: str = None,
+        oauth_type: str = None,
         token_vault_name: str = None,
     ):
         self.callback_url = callback_url
@@ -19,6 +20,7 @@ class CreateOAuth2CredentialProviderShrinkRequest(DaraModel):
         self.description = description
         self.oauth2_credential_provider_name = oauth2_credential_provider_name
         self.oauth2_provider_config_shrink = oauth2_provider_config_shrink
+        self.oauth_type = oauth_type
         self.token_vault_name = token_vault_name
 
     def validate(self):
@@ -44,6 +46,9 @@ class CreateOAuth2CredentialProviderShrinkRequest(DaraModel):
         if self.oauth2_provider_config_shrink is not None:
             result['OAuth2ProviderConfig'] = self.oauth2_provider_config_shrink
 
+        if self.oauth_type is not None:
+            result['OAuthType'] = self.oauth_type
+
         if self.token_vault_name is not None:
             result['TokenVaultName'] = self.token_vault_name
 
@@ -65,6 +70,9 @@ class CreateOAuth2CredentialProviderShrinkRequest(DaraModel):
 
         if m.get('OAuth2ProviderConfig') is not None:
             self.oauth2_provider_config_shrink = m.get('OAuth2ProviderConfig')
+
+        if m.get('OAuthType') is not None:
+            self.oauth_type = m.get('OAuthType')
 
         if m.get('TokenVaultName') is not None:
             self.token_vault_name = m.get('TokenVaultName')